LEGION/res_dev/development/production/generated_projects/final_api_service_project/requirements.txt
requests
pydantic
click
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0
redis>=5.0.0
celery>=5.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
structlog>=23.2.0
